refactor(build): separate deployments for codesets and esco API
