+++ .github/workflows/deploy-dev.yml
@@ -3,6 +3,17 @@ on:
     push:
         branches: ['main']
     workflow_dispatch:
+        inputs:
+            deploy_codesets:
+                description: Deploy codesets
+                type: boolean
+                required: false
+                default: true
+            deploy_esco_api:
+                description: Deploy Esco API
+                type: boolean
+                required: false
+                default: true
 
 jobs:
     deploy-dev:
@@ -10,3 +21,5 @@ jobs:
         secrets: inherit
         with:
             environment: dev
+            deploy_codesets: ${{ github.event.inputs.deploy_codesets }} 
+            deploy_esco_api: ${{ github.event.inputs.deploy_esco_api }}
