+++ infra/apps/codesets/index.ts
@@ -1,14 +1,13 @@
 import * as pulumi from '@pulumi/pulumi';
+import { getSetup } from '../../utils/Setup';
 import {
     createCacheInvalidation,
     createCloudFrontDistribution,
     createOriginAccessIdentity,
 } from './resources/CloudFront';
 import createLambdaAtEdgeFunction from './resources/LambdaAtEdge';
-import createS3Bucket, {createS3BucketPermissions, uploadAssetsToBucket} from './resources/S3Bucket';
-import {getSetup} from './utils/Setup';
-import {createStandardLogsBucket} from "./resources/standardLogsBucket";
-import { createEscoApiLambdaFunctionUrl } from './resources/LambdaFunctionUrl';
+import createS3Bucket, { createS3BucketPermissions, uploadAssetsToBucket } from './resources/S3Bucket';
+import { createStandardLogsBucket } from "./resources/standardLogsBucket";
 
 const setup = getSetup();
 const originAccessIdentity = createOriginAccessIdentity(setup);
@@ -26,17 +25,10 @@ const cloudFrontDistribution = createCloudFrontDistribution(
 uploadAssetsToBucket(s3bucketSetup.bucket);
 createCacheInvalidation(setup, cloudFrontDistribution);
 
-// Codesets
 export const url = pulumi.interpolate`https://${cloudFrontDistribution.domainName}`;
 export const bucketName = s3bucketSetup.bucket.bucket;
-export const lambdaId = pulumi.interpolate`${edgeLambdaPackage.lambdaAtEdgeFunction.name}:${edgeLambdaPackage.lambdaAtEdgeFunction.version}`;
 export const cloudFrontDistributionId = cloudFrontDistribution.id;
 export const standardLogsBucketDetails = {
     arn: standardLogsBucket.arn,
     id: standardLogsBucket.id
 }
-
-// Esco API
-const escoApi = createEscoApiLambdaFunctionUrl(setup, url);
-export const escoApiUrl = escoApi.lambdaFunctionUrl.functionUrl;
-export const escoApiLambdaId = escoApi.lambdaFunction.id;
