+++ infra/apps/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "escoapi-infra",
+  "main": "index.ts",
+  "dependencies": {
+    "@pulumi/aws": "^5.41.0",
+    "@pulumi/pulumi": "^3.69.0"
+  },
+  "scripts": {
+    "prebuild": "rimraf dist",
+    "build": "mkdir -p ./dist && cp -r ../dist/* ./dist/ && cp ../package.json ./dist/ && cp ../package-lock.json ./dist/ && npm install --prefix ./dist --omit=dev",
+    "postbuild": "mv ./dist/node_modules ./dist/codesets/ && rm ./dist/codesets/build/*.js"
+  },
+  "devDependencies": {
+    "@types/mime": "^3.0.1",
+    "rimraf": "^4.1.2"
+  }
+}
