+++ infra/package.json
@@ -1,19 +1,12 @@
 {
-  "name": "codesets-infra",
-  "main": "index.ts",
+  "name": "infra",
   "dependencies": {
     "@pulumi/aws": "^5.41.0",
-    "@pulumi/command": "^0.7.2",
-    "@pulumi/pulumi": "^3.69.0",
-    "mime": "^3.0.0"
+    "@pulumi/pulumi": "^3.69.0"
   },
   "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "mkdir -p ./dist && cp -r ../dist/* ./dist/ && cp ../package.json ./dist/ && cp ../package-lock.json ./dist/ && npm install --prefix ./dist --omit=dev",
-    "postbuild": "mv ./dist/node_modules ./dist/codesets/ && rm ./dist/codesets/build/*.js"
-  },
-  "devDependencies": {
-    "@types/mime": "^3.0.1",
-    "rimraf": "^4.1.2"
+    "build": "npm run build:codesets && npm run build:escoApi",
+    "build:codesets": "npm -C apps/codesets install && npm -C apps/codesets run build",
+    "build:escoApi": "npm -C apps/escoApi install && npm -C apps/escoApi run build"
   }
 }
