+++ package.json
@@ -11,7 +11,9 @@
     "build:infra": "npm -C infra install && npm -C infra run build",
     "clean": "rimraf dist",
     "predeploy": "npm run build && npm run build:infra",
-    "deploy": "pulumi -C infra up --yes",
+    "deploy": "npm run deploy:codesets && npm run deploy:escoapi",
+    "deploy:codesets": "pulumi -C infra/apps/codesets up",
+    "deploy:escoapi": "pulumi -C infra/apps/escoApi up",
     "start": "nodemon --delay 80ms --watch 'src' -e 'ts'  --exec \"npm run build && npm run serve\"",
     "serve": "serverless offline --config ./serverless-offline.yml",
     "test": "jest",
